adding build step
