+++ apps/puppeteer/.gitignore
@@ -1,0 +1,2 @@
+.cache
+dist-e2e
