+++ apps/puppeteer/puppeteer.config.cjs
@@ -1,0 +1,9 @@
+const { join } = require("path");
+
+/**
+ * @type {import("puppeteer").Configuration}
+ */
+module.exports = {
+  // Changes the cache location for Puppeteer.
+  cacheDirectory: join(__dirname, ".cache", "puppeteer"),
+};
