+++ configs/esbuild/index.js
@@ -1,8 +1,10 @@
 import * as esbuild from "esbuild";
 
-export function build(entryPoints, outfile) {
+
+export function build(entryPoints, outfile, external = []) {
   esbuild.build({
     entryPoints,
+    loader: { ".node": "file" },
     bundle: true,
     banner: {
       // js: `#!/usr/bin/env node`,
@@ -10,5 +12,6 @@ export function build(entryPoints, outfile) {
     platform: "node",
     target: "node18",
     outfile,
+    external,
   });
 }
